Dashboard/highlighter/requirements.txt
altair==5.0.1
anyio==3.7.0
attrs==23.1.0
blinker==1.6.2
blis==0.7.9
cachetools==5.3.1
catalogue==2.0.8
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
decorator==5.1.1
en-core-web-lg @ https://github.com/explosion/spacy-models/releases/download/en_core_web_lg-3.5.0/en_core_web_lg-3.5.0-py3-none-any.whl#sha256=c8ac64840c1eb3e3ca7bd38bd1e1c48fb0faeb2449d54d01d5ce629af4595775
exceptiongroup==1.1.1
fastapi==0.98.0
filelock==3.12.2
fsspec==2023.6.0
gensim==4.3.1
gitdb==4.0.10
GitPython==3.1.31
h11==0.14.0
htbuilder==0.6.1
huggingface==0.0.1
huggingface-hub==0.15.1
idna==3.4
importlib-metadata==6.7.0
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
langcodes==3.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
more-itertools==9.1.0
murmurhash==1.0.9
nltk==3.8.1
numpy==1.25.0
packaging==23.1
pandas==2.0.2
pathy==0.10.2
Pillow==9.5.0
preshed==3.0.8
protobuf==4.23.3
pyarrow==12.0.1
pydantic==1.10.9
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.6.3
requests==2.31.0
rich==13.4.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
st-annotated-text==4.0.0
starlette==0.27.0
streamlit==1.23.1
tenacity==8.2.2
thinc==8.1.10
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.6.3
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.3
uvicorn==0.22.0
validators==0.20.0
wasabi==1.1.2
zipp==3.15.0
